Refactor main container to ViewPagerFragment in preparation for using Navigation Arch Component
