+++ app/src/main/java/com/bandonleon/cryptofolio/MainActivity.kt
@@ -14,48 +14,5 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
-        bottomNavigation.setOnNavigationItemSelectedListener listener@ { item ->
-            when (item.itemId) {
-                R.id.navigation_home -> {
-                    displayFragment(PortfolioFragment.newInstance())
-                    return@listener true
-                }
-
-                R.id.navigation_dashboard -> {
-                    displayFragment(PortfolioFragment.newInstance())
-                    return@listener true
-                }
-
-                R.id.navigation_notifications -> {
-                    displayFragment(PortfolioFragment.newInstance())
-                    return@listener true
-                }
-
-                else -> return@listener false
-            }
-        }
-
-        // waitForDebugger()
-
-        thread(start = true) {
-            val coinDatabase = CoinDatabase.getInstance(this)
-            val assetDao = coinDatabase.assetDao()
-            val assets = assetDao.getAll()
-            if (assets == null || assets.isEmpty()) {
-                assetDao.insert(Asset("bitcoin", 3.5f))
-                assetDao.insert(Asset("ethereum", 22.5f))
-            }
-        }
-
-        bottomNavigation.selectedItemId = R.id.navigation_home
-    }
-
-    private fun displayFragment(fragment: Fragment) {
-        val transaction = supportFragmentManager.beginTransaction()
-        transaction.replace(R.id.fragmentContainer, fragment)
-        // Calling addToBackStack() below will allow the use of back button
-        // transaction.addToBackStack(null)
-        transaction.commit()
     }
 }
